grt: move EdgeShift functions to utility
