+++ src/grt/src/fastroute/src/RSMT.cpp
@@ -40,7 +40,6 @@
 
 #include "DataProc.h"
 #include "DataType.h"
-#include "EdgeShift.h"
 #include "RipUp.h"
 #include "flute.h"
 #include "pdr/pdrev.h"
